refactor: split tree_tests.rs to stay under 1000-line limit
Split iterative tree tests into tests/iterative_tree_tests.rs (650 lines)
and kept recursive tree tests in tests/tree_tests.rs (379 lines).
